Move test utils into tests
