+++ nlp_primitives/tests/test_mean_characters_per_word.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 
 from ..mean_characters_per_word import MeanCharactersPerWord
-from ..utils import PrimitiveT, find_applicable_primitives, valid_dfs
+from .test_utils import PrimitiveT, find_applicable_primitives, valid_dfs
 
 
 class TestMeanCharactersPerWord(PrimitiveT):
