+++ nlp_primitives/tests/test_part_of_speech_count.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 
 from ..part_of_speech_count import PartOfSpeechCount
-from ..utils import PrimitiveT, find_applicable_primitives, valid_dfs
+from .test_utils import PrimitiveT, find_applicable_primitives, valid_dfs
 
 
 class TestPartOfSpeechCount(PrimitiveT):
