+++ nlp_primitives/tests/test_polarity_score.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 
 from ..polarity_score import PolarityScore
-from ..utils import PrimitiveT, find_applicable_primitives, valid_dfs
+from .test_utils import PrimitiveT, find_applicable_primitives, valid_dfs
 
 
 class TestPolarityScore(PrimitiveT):
