+++ nlp_primitives/tests/test_title_word_count.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 
 from ..title_word_count import TitleWordCount
-from ..utils import PrimitiveT, find_applicable_primitives, valid_dfs
+from .test_utils import PrimitiveT, find_applicable_primitives, valid_dfs
 
 
 class TestTitleWordCount(PrimitiveT):
